sqlite: unify benchmarks
